move AuthLayout.jsx to components/auth
